requirements.txt
absl-py==1.4.0
accelerate==0.21.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autoattack @ git+https://github.com/fra31/auto-attack@a39220048b3c9f2cca9a4d3a54604793c68eca7e
Babel==2.14.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==6.0.0
blessed==1.20.0
blosc2==2.4.0
cachetools==5.3.1
certifi==2022.6.15.1
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.3
cmake==3.27.1
comm==0.2.1
cox==0.1.post3
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
docker-pycreds==0.4.0
einops==0.4.1
entrypoints==0.3
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.12.2
flake8==3.7.9
fonttools==4.37.1
fqdn==1.5.1
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gdown==4.7.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gpustat==1.1.1
GPUtil==1.4.0
grpcio==1.56.2
h11==0.14.0
h5py==3.7.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.16.4
idna==3.3
importlib-metadata==6.8.0
ipykernel==6.29.3
ipython==8.22.1
ipywidgets==8.1.2
isoduration==20.11.0
isort==4.3.21
jedi==0.19.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kaggle==1.5.16
kiwisolver==1.4.4
kornia==0.6.8
lit==16.0.6
littleutils==0.2.2
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.7
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
ndindex==1.7
nest-asyncio==1.6.0
networkx==3.1
notebook==7.1.1
notebook_shim==0.2.4
numexpr==2.8.8
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
ogb==1.3.6
opencv-python==4.9.0.80
outdated==0.2.2
overrides==7.7.0
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pathtools==0.1.2
patsy==0.5.3
pexpect==4.9.0
Pillow==9.2.0
platformdirs==4.2.0
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.1
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.5.0
pycountry==22.3.5
pycparser==2.21
pyflakes==2.1.1
Pygments==2.17.2
pyparsing==3.0.9
pyre-extensions==0.0.29
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytorch-ignite==0.4.12
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
regex==2022.8.17
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
robustness==1.2.1.post2
rpds-py==0.18.0
rsa==4.9
safetensors==0.3.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.13.1
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.9.8
setGPU==0.0.7
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.4.1
stack-data==0.6.3
statsmodels==0.14.0
sympy==1.12
tables==3.9.2
tabulate==0.9.0
tb-nightly==2.14.0a20230806
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
terminado==0.18.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.64.1
traitlets==5.14.1
transformers==4.31.0
triton==2.0.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.3.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.12
wandb==0.13.3
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.3.6
widgetsnbextension==4.0.10
wilds==1.2.2
xformers==0.0.20
xmltodict==0.13.0
yacs==0.1.8
yapf==0.29.0
zipp==3.16.2
